+++ srcs/user/User.cpp
@@ -35,7 +35,7 @@ std::string					User::getNickname(void) const { return this->_nickname; }
 std::string					User::getUsername(void) const { return this->_username; }
 std::string					User::getFullname(void) const { return this->_fullname; }
 std::string					User::getHostname(void) const { return this->_hostname; }
-uint8_t					    User::getMode(void) const { return this->_mode; }
+short					    User::getMode(void) const { return this->_mode; }
 bool 						User::getPassword(void) const { return this->_password; }
 bool 						User::getAuthenticated(void) const {
 	return this->_authenticated;
@@ -46,7 +46,7 @@ std::deque<std::string>	User::getChannelsJoined(void) const {
 int							User::getStatus(void) const { return this->_status; }
 time_t  					User::getLastActivityTime(void) const { return this->_lastActivityTime; }
 time_t  					User::getPingTime(void) const { return this->_pingTime; }
-bool                        User::hasMode(uint8_t mode) { return ((this->_mode & mode) > 0); }
+bool                        User::hasMode(short mode) { return ((this->_mode & mode) > 0); }
 
 // SETTERS
 void User::setNickname(std::string nickname) { this->_nickname = nickname; }
@@ -60,8 +60,8 @@ void User::setAuthenticated(bool authenticated) {
 void User::setStatus(int status) { this->_status = status; }
 void User::setLastActivityTime(void) { this->_lastActivityTime = time(NULL); }
 void User::setPingTime(void) { this->_pingTime = time(NULL); }
-void User::addMode(uint8_t mode) { this->_mode |= mode; }
-void User::removeMode(uint8_t mode) { this->_mode &= ~mode; }
+void User::addMode(short mode) { this->_mode |= mode; }
+void User::removeMode(short mode) { this->_mode &= ~mode; }
 
 // CHANNEL JOINED MANAGEMENT
 bool User::addChannelJoined(std::string channelName) {
